feat-#79: Add game mode with one control (fix)
